final touch: move to ExprAnalyzer
